tests loading properly
