--- apps/prairielearn/elements/pl-big-operator-input/tests/helpers.py
@@ -1,21 +1,0 @@
-from __future__ import annotations
-
-
-class RegressionTestSuite:
-    pass
-
-
-class SmokeTestSuite:
-    pass
-
-
-class UnitTestSuite:
-    pass
-
-
-class AdapterTestSuite:
-    pass
-
-
-class BrowserTestSuite:
-    pass
